refactor: Move typed class and interface into separate files
